Update and thin files from Alamofire
- Also renamed the error type to address PR feedback
